refactor: エクスポートしている型定義を別ファイルに切り出した
