+++ src/components/contributions/by-repo/comments.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { Summary } from '../contributions-by-repo';
+import { toHtmlUrl } from '../contributions';
 import {
-  toHtmlUrl,
   PullRequestReviewCommentEventPayload,
   IssueCommentEventPayload,
   PullRequestReviewEventPayload,
   CommitCommentEventPayload,
-} from '../contributions';
+  Summary,
+} from '../types';
 import {
   CommentIcon,
   CommitIcon,
